Fix test failed.
