+++ microproxy/test/layer/test_forward.py
@@ -5,7 +5,7 @@ from tornado.locks import Semaphore
 from tornado.iostream import IOStream
 from tornado.netutil import add_accept_handler
 
-from microproxy.context import Context
+from microproxy.context import LayerContext
 from microproxy.layer import ForwardLayer
 
 
@@ -35,8 +35,8 @@ class ForwardLayerTest(AsyncTestCase):
         self.io_loop.remove_handler(listener)
         listener.close()
 
-        self.context = Context(src_stream=server_streams[0],
-                               dest_stream=client_streams[1])
+        self.context = LayerContext(src_stream=server_streams[0],
+                                    dest_stream=client_streams[1])
         self.src_stream = client_streams[0]
         self.dest_stream = server_streams[1]
         self.forward_layer = ForwardLayer(self.context)
