+++ microproxy/test/run_tests.py
@@ -2,9 +2,9 @@ import unittest
 import tornado.testing
 
 TEST_MODULES = [
-    "microproxy.test.test_http",
     "microproxy.test.test_config",
     "microproxy.test.test_proxy",
+    "microproxy.test.context.test_http",
     "microproxy.test.layer.test_forward",
     "microproxy.test.layer.test_http1",
     "microproxy.test.layer.test_socks",
